Hardware Version 2

+++ MikroC/libPing.h
@@ -15,6 +15,6 @@ int ReadPing_L();
 int ReadPing_F();
 int ReadPing_R();
 
-int Count_Distance(int cnt, int max, int scale);
+int Count_Distance(double cnt, double max, double scale);
 
 #endif
